--- libs/wire-subsystems/src/Wire/ConversationsSubsystem.hs
@@ -1,12 +1,0 @@
-{-# LANGUAGE TemplateHaskell #-}
-
-module Wire.ConversationsSubsystem where
-
-import Data.Id
-import Polysemy
-import Wire.API.Team.Conversation (LeavingConversations)
-
-data ConversationsSubsystem m a where
-  InternalLeavingConversationsFrom :: TeamId -> UserId -> ConversationsSubsystem m LeavingConversations
-
-makeSem ''ConversationsSubsystem
