+++ libs/wire-subsystems/src/Wire/GalleyAPIAccess/Rpc.hs
@@ -45,7 +45,6 @@ import Wire.API.Routes.Internal.Brig.EJPD (EJPDConvInfo)
 import Wire.API.Routes.Internal.Galley.TeamsIntra qualified as Team
 import Wire.API.Routes.Version
 import Wire.API.Team
-import Wire.API.Team.Conversation (LeavingConversations)
 import Wire.API.Team.Conversation qualified as Conv
 import Wire.API.Team.Feature
 import Wire.API.Team.LegalHold
@@ -77,7 +76,6 @@ interpretGalleyAPIAccessToRpc disabledVersions galleyEndpoint =
           NewClient id' ci -> newClient id' ci
           CheckUserCanJoinTeam id' -> checkUserCanJoinTeam id'
           AddTeamMember id' id'' a b -> addTeamMember id' id'' a b
-          RemoveTeamMember zUser' user team -> removeTeamMember zUser' user team
           CreateTeam id' bnt id'' -> createTeam id' bnt id''
           GetTeamMember id' id'' -> getTeamMember id' id''
           GetTeamMembers tid maxResults -> getTeamMembers tid maxResults
@@ -98,7 +96,6 @@ interpretGalleyAPIAccessToRpc disabledVersions galleyEndpoint =
           UnblockConversation lusr mconn qcnv -> unblockConversation v lusr mconn qcnv
           GetEJPDConvInfo uid -> getEJPDConvInfo uid
           GetTeamAdmins tid -> getTeamAdmins tid
-          LeavingConversationsFrom tid uid -> leavingConversationsFrom tid uid
 
 getUserLegalholdStatus ::
   ( Member TinyLog r,
@@ -284,29 +281,6 @@ addTeamMember u tid minvmeta role = do
         . zUser u
         . expect [status200, status403]
         . lbytes (encode bdy)
-
--- | Calls 'Galley.API.uncheckedRemoveTeamMemberH'.
-removeTeamMember ::
-  ( Member Rpc r,
-    Member (Input Endpoint) r,
-    Member TinyLog r
-  ) =>
-  Local UserId ->
-  UserId ->
-  TeamId ->
-  Sem r ()
-removeTeamMember _puid tuid tid = do
-  debug $
-    remote "galley"
-      . msg (val "Removing member from team")
-  void $ galleyRequest req
-  where
-    req =
-      method DELETE
-        . paths ["i", "teams", toByteString' tid, "members"]
-        . header "Content-Type" "application/json"
-        . zUser tuid
-        . expect [status200, status403]
 
 -- | Calls 'Galley.API.createBindingTeamH'.
 createTeam ::
@@ -706,23 +680,3 @@ getEJPDConvInfo uid = do
     getReq =
       method GET
         . paths ["i", "user", toByteString' uid, "all-conversations"]
-
--- | Calls 'Galley.API.updateTeamStatusH'.
-leavingConversationsFrom ::
-  ( Member (Error ParseException) r,
-    Member Rpc r,
-    Member (Input Endpoint) r,
-    Member TinyLog r
-  ) =>
-  TeamId ->
-  UserId ->
-  Sem r LeavingConversations
-leavingConversationsFrom tid uid = do
-  debug $ remote "galley" . msg (val "Leave all conversations of a user in a team")
-  decodeBodyOrThrow "galley" =<< galleyRequest req
-  where
-    req =
-      method POST
-        . paths ["i", "teams", toByteString' tid, "leave-conversations-from", toByteString' uid]
-        . header "Content-Type" "application/json"
-        . expect2xx
