+++ libs/wire-subsystems/wire-subsystems.cabal
@@ -186,8 +186,6 @@ library
     Wire.BlockListStore.Cassandra
     Wire.BrigAPIAccess
     Wire.BrigAPIAccess.Rpc
-    Wire.ConversationsSubsystem
-    Wire.ConversationsSubsystem.GalleyAPI
     Wire.ConversationStore
     Wire.ConversationStore.Cassandra
     Wire.ConversationStore.Cassandra.Instances
