+++ services/brig/src/Brig/CanonicalInterpreter.hs
@@ -47,8 +47,6 @@ import Wire.AuthenticationSubsystem.Config
 import Wire.AuthenticationSubsystem.Interpreter
 import Wire.BlockListStore
 import Wire.BlockListStore.Cassandra
-import Wire.ConversationsSubsystem (ConversationsSubsystem)
-import Wire.ConversationsSubsystem.GalleyAPI (interpretConversationsSubsystemToGalleyAPI)
 import Wire.DeleteQueue
 import Wire.DomainRegistrationStore
 import Wire.DomainRegistrationStore.Cassandra
@@ -177,7 +175,6 @@ type BrigLowerLevelEffects =
      HashPassword,
      UserKeyStore,
      UserStore,
-     ConversationsSubsystem,
      IndexedUserStore,
      SessionStore,
      PasswordStore,
@@ -344,7 +341,6 @@ runBrigToIO e (AppT ma) = do
               . interpretPasswordStore e.casClient
               . interpretSessionStoreCassandra e.casClient
               . interpretIndexedUserStoreES indexedUserStoreConfig
-              . interpretConversationsSubsystemToGalleyAPI
               . interpretUserStoreCassandra e.casClient
               . interpretUserKeyStoreCassandra e.casClient
               . runHashPassword e.settings.passwordHashingOptions
