+++ services/galley/src/Galley/API/Public/TeamMember.hs
@@ -33,3 +33,4 @@ teamMemberAPI =
     <@> mkNamedAPI @"delete-non-binding-team-member" deleteNonBindingTeamMember
     <@> mkNamedAPI @"update-team-member" updateTeamMember
     <@> mkNamedAPI @"get-team-members-csv" Export.getTeamMembersCSV
+    <@> mkNamedAPI @"remove-team-collaborator" removeTeamCollaborator
