+++ services/galley/src/Galley/API/Teams.hs
@@ -52,6 +52,8 @@ module Galley.API.Teams
     ensureNotTooLargeForLegalHold,
     ensureNotTooLargeToActivateLegalHold,
     internalDeleteBindingTeam,
+    removeTeamCollaborator,
+    uncheckedLeaveTeams,
   )
 where
 
@@ -76,14 +78,18 @@ import Data.Set qualified as Set
 import Data.Singletons
 import Data.Time.Clock (UTCTime)
 import Galley.API.Action
+import Galley.API.Cells (shouldPushToCells)
 import Galley.API.Error as Galley
 import Galley.API.LegalHold.Team
+import Galley.API.MLS.Removal (RemoveUserIncludeMain (..), removeUser)
+import Galley.API.Query qualified as Query
 import Galley.API.Teams.Features.Get
 import Galley.API.Teams.Notifications qualified as APITeamQueue
 import Galley.API.Update qualified as API
 import Galley.API.Util
 import Galley.App
 import Galley.Effects
+import Galley.Effects.BackendNotificationQueueAccess (enqueueNotification)
 import Galley.Effects.ExternalAccess qualified as E
 import Galley.Effects.LegalHoldStore qualified as Data
 import Galley.Effects.Queue qualified as E
@@ -95,22 +101,34 @@ import Galley.Intra.Journal qualified as Journal
 import Galley.Options
 import Galley.Types.Teams
 import Imports hiding (forkIO)
+import Network.AMQP qualified as Q
 import Polysemy
 import Polysemy.Error
 import Polysemy.Input
 import Polysemy.TinyLog qualified as P
 import System.Logger qualified as Log
-import Wire.API.Conversation (ConversationRemoveMembers (..))
+import Wire.API.Conversation (ConvIdsPage, ConvType (..), ConversationRemoveMembers (..), pattern GetPaginatedConversationIds)
+import Wire.API.Conversation.Action
+  ( SomeConversationAction (..),
+  )
 import Wire.API.Conversation.Role (wireConvRoles)
 import Wire.API.Conversation.Role qualified as Public
 import Wire.API.Error
 import Wire.API.Error.Galley
+import Wire.API.Event.Conversation (EventData (EdMembersLeave))
 import Wire.API.Event.Conversation qualified as Conv
 import Wire.API.Event.LeaveReason
 import Wire.API.Event.Team
+import Wire.API.Federation.API
+  ( fedQueueClient,
+    sendBundle,
+  )
+import Wire.API.Federation.API.Galley.Notifications (ConversationUpdate (..), GalleyNotificationTag (OnUserDeletedConversationsTag), UserDeletedConversationsNotification (..), UserDeletedNotificationMaxConvs)
+import Wire.API.Federation.API.Util (makeConversationUpdateBundle)
 import Wire.API.Federation.Error
+import Wire.API.Push.V2 qualified as PushV2
 import Wire.API.Routes.Internal.Galley.TeamsIntra
-import Wire.API.Routes.MultiTablePaging (MultiTablePage (MultiTablePage), MultiTablePagingState (mtpsState))
+import Wire.API.Routes.MultiTablePaging (MultiTablePage (..), MultiTablePagingState (mtpsState))
 import Wire.API.Routes.Public.Galley.TeamMember
 import Wire.API.Team
 import Wire.API.Team qualified as Public
@@ -134,6 +152,7 @@ import Wire.Sem.Now
 import Wire.Sem.Now qualified as Now
 import Wire.Sem.Paging.Cassandra
 import Wire.StoredConversation
+import Wire.StoredConversation qualified as Data
 import Wire.TeamCollaboratorsSubsystem
 import Wire.TeamSubsystem (TeamSubsystem)
 import Wire.TeamSubsystem qualified as TeamSubsystem
@@ -1297,3 +1316,185 @@ checkAdminLimit :: (Member (ErrorS 'TooManyTeamAdmins) r) => Int -> Sem r ()
 checkAdminLimit adminCount =
   when (adminCount > 2000) $
     throwS @'TooManyTeamAdmins
+
+-- | Removing a team collaborator and clean their conversations
+removeTeamCollaborator ::
+  forall p1 r.
+  ( p1 ~ CassandraPaging,
+    Member BackendNotificationQueueAccess r,
+    Member ConversationStore r,
+    Member (Error DynError) r,
+    Member (Error FederationError) r,
+    Member (Error InternalError) r,
+    Member (ErrorS OperationDenied) r,
+    Member (ErrorS NotATeamMember) r,
+    Member ExternalAccess r,
+    Member NotificationSubsystem r,
+    Member (Input Env) r,
+    Member (Input Opts) r,
+    Member Now r,
+    Member (ListItems p1 ConvId) r,
+    Member (ListItems p1 (Remote ConvId)) r,
+    Member ProposalStore r,
+    Member P.TinyLog r,
+    Member Random r,
+    Member TeamFeatureStore r,
+    Member TeamStore r,
+    Member TeamCollaboratorsSubsystem r,
+    Member BrigAPIAccess r
+  ) =>
+  Local UserId ->
+  TeamId ->
+  UserId ->
+  Sem r ()
+removeTeamCollaborator lusr tid rusr = do
+  P.debug $
+    Log.field "targets" (toByteString rusr)
+      . Log.field "action" (Log.val "Teams.removeTeamCollaborator")
+  zusrMember <- E.getTeamMember tid (tUnqualified lusr)
+  void $ permissionCheck RemoveTeamMember zusrMember
+  uncheckedLeaveTeams (lusr $> rusr) Nothing [tid]
+
+uncheckedLeaveTeams ::
+  forall p1 r.
+  ( p1 ~ CassandraPaging,
+    Member BackendNotificationQueueAccess r,
+    Member ConversationStore r,
+    Member (Error DynError) r,
+    Member (Error FederationError) r,
+    Member (Error InternalError) r,
+    Member ExternalAccess r,
+    Member NotificationSubsystem r,
+    Member (Input Env) r,
+    Member (Input Opts) r,
+    Member Now r,
+    Member (ListItems p1 ConvId) r,
+    Member (ListItems p1 (Remote ConvId)) r,
+    Member ProposalStore r,
+    Member P.TinyLog r,
+    Member Random r,
+    Member TeamFeatureStore r,
+    Member TeamStore r,
+    Member TeamCollaboratorsSubsystem r,
+    Member BrigAPIAccess r
+  ) =>
+  Local UserId ->
+  Maybe ConnId ->
+  [TeamId] ->
+  Sem r ()
+uncheckedLeaveTeams lusr conn tids = do
+  let nRange1000 = toRange (Proxy @1000) :: Range 1 1000 Int32
+  forM_ tids $ \tid -> do
+    toNotify <-
+      handleImpossibleErrors $
+        getFeatureForTeam @LimitedEventFanoutConfig tid
+          >>= ( \case
+                  FeatureStatusEnabled -> Left <$> E.getTeamAdmins tid
+                  FeatureStatusDisabled -> Right <$> getTeamMembersForFanout tid
+              )
+            . (.status)
+    uncheckedDeleteTeamMember lusr conn tid (tUnqualified lusr) toNotify
+
+  allConvIds <- Query.conversationIdsPageFrom lusr (GetPaginatedConversationIds Nothing nRange1000)
+  contacts <- E.getContactList (tUnqualified lusr)
+  goConvPages contacts nRange1000 allConvIds
+  where
+    goConvPages :: [UserId] -> Range 1 1000 Int32 -> ConvIdsPage -> Sem r ()
+    goConvPages contacts range page = do
+      let (localConvs, remoteConvs) = partitionQualified lusr (mtpResults page)
+      leaveLocalConversations contacts localConvs
+      traverse_ leaveRemoteConversations (rangedChunks remoteConvs)
+      when (mtpHasMore page) $ do
+        let nextState = mtpPagingState page
+            nextQuery = GetPaginatedConversationIds (Just nextState) range
+        newCids <- Query.conversationIdsPageFrom lusr nextQuery
+        goConvPages contacts range newCids
+
+    -- The @'NotATeamMember@ and @'TeamNotFound@ errors cannot happen at this
+    -- point: the user is a team member because we fetched the list of teams
+    -- they are member of, and conversely the list of teams was fetched exactly
+    -- for this user so it cannot be that the team is not found. Therefore, this
+    -- helper just drops the errors.
+    handleImpossibleErrors ::
+      Sem
+        ( ErrorS 'NotATeamMember
+            ': ErrorS 'TeamNotFound
+            ': r
+        )
+        a ->
+      Sem r a
+    handleImpossibleErrors action =
+      mapToDynamicError @'TeamNotFound (mapToDynamicError @'NotATeamMember action)
+
+    leaveLocalConversations :: [UserId] -> [ConvId] -> Sem r ()
+    leaveLocalConversations contacts ids = do
+      let qUser = tUntagged lusr
+      cc <- E.getConversations ids
+      now <- Now.get
+      pp <- for cc $ \c -> case Data.convType c of
+        SelfConv -> pure Nothing
+        One2OneConv -> do
+          localMembers <- E.getLocalMembers c.id_
+          if any (flip notElem (tUnqualified lusr : contacts) . (.id_)) localMembers
+            then E.deleteMembers c.id_ (UserList [tUnqualified lusr] []) $> Nothing
+            else E.deleteConversation c.id_ $> Nothing
+        ConnectConv -> E.deleteMembers c.id_ (UserList [tUnqualified lusr] []) $> Nothing
+        RegularConv
+          | tUnqualified lusr `isMember` c.localMembers -> do
+              runError (removeUser (qualifyAs lusr c) RemoveUserIncludeMain (tUntagged lusr)) >>= \case
+                Left e -> P.err $ Log.msg ("failed to send remove proposal: " <> internalErrorDescription e)
+                Right _ -> pure ()
+              E.deleteMembers c.id_ (UserList [tUnqualified lusr] [])
+              Just <$> notifyRemoteMembersAndPrepareLocalMembersLeft now qUser c
+          | otherwise -> pure Nothing
+
+      pushNotifications (catMaybes pp)
+
+    notifyRemoteMembersAndPrepareLocalMembersLeft :: UTCTime -> Qualified UserId -> StoredConversation -> Sem r Push
+    notifyRemoteMembersAndPrepareLocalMembersLeft now qUser c = do
+      let e =
+            Conv.Event
+              { evtConv = tUntagged (qualifyAs lusr c.id_),
+                evtSubConv = Nothing,
+                evtFrom = tUntagged lusr,
+                evtTime = now,
+                evtTeam = Nothing,
+                evtData = EdMembersLeave EdReasonDeleted (U.QualifiedUserIdList [qUser])
+              }
+      for_ (bucketRemote (fmap (.id_) c.remoteMembers)) $ notifyRemoteMembers now qUser c
+      pure $
+        def
+          { origin = Just (tUnqualified lusr),
+            json = toJSONObject e,
+            recipients = map localMemberToRecipient c.localMembers,
+            isCellsEvent = shouldPushToCells c.metadata e,
+            conn,
+            route = PushV2.RouteDirect
+          }
+
+    -- FUTUREWORK: This could be optimized to reduce the number of RPCs
+    -- made. When a team is deleted the burst of RPCs created here could
+    -- lead to performance issues. We should cover this in a performance
+    -- test.
+    notifyRemoteMembers :: UTCTime -> Qualified UserId -> StoredConversation -> Remote [UserId] -> Sem r ()
+    notifyRemoteMembers now qUser c remotes = do
+      let cid = c.id_
+          convUpdate =
+            ConversationUpdate
+              { time = now,
+                origUserId = qUser,
+                convId = cid,
+                alreadyPresentUsers = tUnqualified remotes,
+                action = SomeConversationAction (sing @'ConversationLeaveTag) (),
+                extraConversationData = def
+              }
+      enqueueNotification Q.Persistent remotes $ do
+        makeConversationUpdateBundle convUpdate
+          >>= sendBundle
+
+    leaveRemoteConversations :: Range 1 UserDeletedNotificationMaxConvs [Remote ConvId] -> Sem r ()
+    leaveRemoteConversations cids =
+      for_ (bucketRemote (fromRange cids)) $ \remoteConvs -> do
+        let userDelete = UserDeletedConversationsNotification (tUnqualified lusr) (unsafeRange (tUnqualified remoteConvs))
+        let rpc = fedQueueClient @'OnUserDeletedConversationsTag userDelete
+        enqueueNotification Q.Persistent remoteConvs rpc
